combining two ROW tasks
